+++ app/Http/Middleware/VerifyCsrfToken.php
@@ -2,9 +2,9 @@
 
 namespace App\Http\Middleware;
 
-use Illuminate\Foundation\Http\Middleware\VerifyCsrfToken as Middleware;
+use Illuminate\Foundation\Http\Middleware\PreventRequestForgery;
 
-class VerifyCsrfToken extends Middleware
+class VerifyCsrfToken extends PreventRequestForgery
 {
     /**
      * The URIs that should be excluded from CSRF verification.
