+++ tests/Feature/ArtikelUploadTest.php
@@ -14,7 +14,7 @@ class ArtikelUploadTest extends BaseTestCase
         Storage::fake('public');
     }
 
-    /** @test */
+    #[Test]
     public function it_can_upload_valid_image()
     {
         $file = UploadedFile::fake()->image('test-artikel.jpg', 800, 600);
@@ -34,7 +34,7 @@ class ArtikelUploadTest extends BaseTestCase
         ]);
     }
 
-    /** @test */
+    #[Test]
     public function it_rejects_non_image_files()
     {
         $file = UploadedFile::fake()->create('document.pdf', 100);
@@ -47,7 +47,7 @@ class ArtikelUploadTest extends BaseTestCase
         $this->assertNotEquals(200, $response->status());
     }
 
-    /** @test */
+    #[Test]
     public function it_rejects_files_larger_than_2mb()
     {
         $file = UploadedFile::fake()->image('large-image.jpg')->size(3000); // 3MB
@@ -60,7 +60,7 @@ class ArtikelUploadTest extends BaseTestCase
         $this->assertNotEquals(200, $response->status());
     }
 
-    /** @test */
+    #[Test]
     public function it_accepts_jpg_format()
     {
         $file = UploadedFile::fake()->image('test.jpg');
@@ -73,7 +73,7 @@ class ArtikelUploadTest extends BaseTestCase
         $response->assertJsonFragment(['success' => true]);
     }
 
-    /** @test */
+    #[Test]
     public function it_accepts_jpeg_format()
     {
         $file = UploadedFile::fake()->image('test.jpeg');
@@ -86,7 +86,7 @@ class ArtikelUploadTest extends BaseTestCase
         $response->assertJsonFragment(['success' => true]);
     }
 
-    /** @test */
+    #[Test]
     public function it_accepts_png_format()
     {
         $file = UploadedFile::fake()->image('test.png');
@@ -99,7 +99,7 @@ class ArtikelUploadTest extends BaseTestCase
         $response->assertJsonFragment(['success' => true]);
     }
 
-    /** @test */
+    #[Test]
     public function it_requires_file_parameter()
     {
         $response = $this->post(route('artikel.upload_gambar'), []);
@@ -108,7 +108,7 @@ class ArtikelUploadTest extends BaseTestCase
         $this->assertContains($response->status(), [302, 400, 500]);
     }
 
-    /** @test */
+    #[Test]
     public function upload_returns_valid_url()
     {
         $file = UploadedFile::fake()->image('artikel-gambar.jpg');
@@ -125,7 +125,7 @@ class ArtikelUploadTest extends BaseTestCase
         $this->assertNotEmpty($data['url']);
     }
 
-    /** @test */
+    #[Test]
     public function upload_returns_valid_path()
     {
         $file = UploadedFile::fake()->image('artikel-gambar.jpg');
@@ -142,7 +142,7 @@ class ArtikelUploadTest extends BaseTestCase
         $this->assertStringContainsString('uploads/artikel', $data['path']);
     }
 
-    /** @test */
+    #[Test]
     public function upload_requires_authentication()
     {
         auth()->logout();
