+++ tests/Feature/DasborControllerTest.php
@@ -10,7 +10,7 @@ class DasborControllerTest extends BaseTestCase
 {
     use DatabaseTransactions;
 
-    /** @test */
+    #[Test]
     public function it_can_access_dasbor_index()
     {
         $response = $this->get(route('dasbor'));
@@ -21,7 +21,7 @@ class DasborControllerTest extends BaseTestCase
         $response->assertViewHas('categoriesItems');
     }
 
-    /** @test */
+    #[Test]
     public function it_returns_correct_categories_items_structure()
     {
         $response = $this->get(route('dasbor'));
@@ -35,7 +35,7 @@ class DasborControllerTest extends BaseTestCase
         });
     }
 
-    /** @test */
+    #[Test]
     public function it_has_correct_category_keys()
     {
         $response = $this->get(route('dasbor'));
