+++ tests/Feature/DataPresisiExcelDownloadTest.php
@@ -22,7 +22,7 @@ class DataPresisiExcelDownloadTest extends BaseTestCase
         ],
     ];
 
-    /** @test */
+    #[Test]
     public function test_excel_download_button_exists_in_pangan_page()
     {
         $response = $this->get('/data-presisi/pangan');
@@ -40,7 +40,7 @@ class DataPresisiExcelDownloadTest extends BaseTestCase
         }
     }
 
-    /** @test */
+    #[Test]
     public function test_excel_download_button_exists_in_pendidikan_page()
     {
         $response = $this->get('/data-presisi/pendidikan');
@@ -58,7 +58,7 @@ class DataPresisiExcelDownloadTest extends BaseTestCase
         }
     }
 
-    /** @test */
+    #[Test]
     public function test_excel_download_button_component_renders_correctly()
     {
         $successCount = 0;
@@ -87,7 +87,7 @@ class DataPresisiExcelDownloadTest extends BaseTestCase
         );
     }
 
-    /** @test */
+    #[Test]
     public function test_pangan_and_pendidikan_have_matching_excel_button_structure()
     {
         $panganResponse = $this->get('/data-presisi/pangan');
