+++ tests/Feature/DataPresisiFilterTahunJavaScriptTest.php
@@ -21,7 +21,7 @@ class DataPresisiFilterTahunJavaScriptTest extends BaseTestCase
         '/dtks/sandang'
     ];
 
-    /** @test */
+    #[Test]
     public function test_filter_tahun_javascript_event_listener_exists_in_all_modules()
     {
         $successCount = 0;
@@ -60,7 +60,7 @@ class DataPresisiFilterTahunJavaScriptTest extends BaseTestCase
         }
     }
 
-    /** @test */
+    #[Test]
     public function test_filter_tahun_ajax_parameter_exists_in_all_modules()
     {
         $successCount = 0;
@@ -97,7 +97,7 @@ class DataPresisiFilterTahunJavaScriptTest extends BaseTestCase
         }
     }
 
-    /** @test */
+    #[Test]
     public function test_datatable_reload_functionality_exists()
     {
         $successCount = 0;
@@ -134,7 +134,7 @@ class DataPresisiFilterTahunJavaScriptTest extends BaseTestCase
         }
     }
 
-    /** @test */
+    #[Test]
     public function test_chart_reload_functionality_exists()
     {
         $successCount = 0;
