+++ tests/Feature/GlobalRateLimiterTest.php
@@ -19,7 +19,7 @@ class GlobalRateLimiterTest extends TestCase
         Cache::flush();
     }
 
-    /** @test */
+    #[Test]
     public function it_allows_requests_when_rate_limiter_is_disabled()
     {
         // Disable rate limiter
@@ -33,7 +33,7 @@ class GlobalRateLimiterTest extends TestCase
         }
     }
 
-    /** @test */
+    #[Test]
     public function it_allows_requests_within_limit_when_rate_limiter_is_enabled()
     {
         // Enable rate limiter with low limits for testing
@@ -52,7 +52,7 @@ class GlobalRateLimiterTest extends TestCase
         }
     }
 
-    /** @test */
+    #[Test]
     public function it_blocks_requests_when_limit_is_exceeded()
     {
         // Enable rate limiter with low limits for testing
@@ -72,7 +72,7 @@ class GlobalRateLimiterTest extends TestCase
         $this->assertEquals(429, $response3->getStatusCode());
     }
 
-    /** @test */
+    #[Test]
     public function it_returns_correct_json_response_for_api_requests()
     {
         // Enable rate limiter with low limits for testing
@@ -95,7 +95,7 @@ class GlobalRateLimiterTest extends TestCase
         $response->assertJsonStructure(['retry_after']);
     }
 
-    /** @test */
+    #[Test]
     public function it_excludes_configured_paths_from_rate_limiting()
     {
         // Enable rate limiter with low limits
@@ -114,7 +114,7 @@ class GlobalRateLimiterTest extends TestCase
         }
     }
 
-    /** @test */
+    #[Test]
     public function it_excludes_configured_ip_addresses_from_rate_limiting()
     {
         // Enable rate limiter with low limits
@@ -133,7 +133,7 @@ class GlobalRateLimiterTest extends TestCase
         }
     }
 
-    /** @test */
+    #[Test]
     public function it_respects_wildcard_patterns_in_excluded_paths()
     {
         // Enable rate limiter with low limits
@@ -155,7 +155,7 @@ class GlobalRateLimiterTest extends TestCase
         $this->assertNotEquals(429, $response3->getStatusCode());
     }
 
-    /** @test */
+    #[Test]
     public function it_respects_different_rate_limits_for_different_ips()
     {
         // Enable rate limiter with low limits
@@ -176,7 +176,7 @@ class GlobalRateLimiterTest extends TestCase
         $this->assertNotEquals(429, $response->getStatusCode());
     }
 
-    /** @test */
+    #[Test]
     public function it_works_with_simple_paths()
     {
         // Enable rate limiter with low limits
