+++ tests/Feature/MenuControllerCmsTest.php
@@ -11,9 +11,7 @@ class MenuControllerCmsTest extends BaseTestCase
 {
     use DatabaseTransactions;
 
-    /** @test
-     * Menguji halaman index menu dapat ditampilkan tanpa error
-     */
+    #[Test]
     public function test_index_menampilkan_halaman_daftar_menu()
     {
         // Arrange: siapkan page dan category
@@ -30,9 +28,7 @@ class MenuControllerCmsTest extends BaseTestCase
         $response->assertViewHas('sourceItem');
     }
 
-    /** @test
-     * Menguji penyimpanan data menu baru berhasil dan redirect sesuai jenis menu
-     */
+    #[Test]
     public function test_store_menyimpan_menu_baru_dan_redirect_sesuai_jenis()
     {
         // Arrange
@@ -57,9 +53,7 @@ class MenuControllerCmsTest extends BaseTestCase
         $this->assertEquals('Menu berhasil disimpan.', session('success'));
     }
 
-    /** @test
-     * Menguji redirect store jika tipe menu 2
-     */
+    #[Test]
     public function test_store_redirect_ke_index_dengan_type_dua()
     {
         $payload = [
@@ -80,9 +74,7 @@ class MenuControllerCmsTest extends BaseTestCase
         $this->assertDatabaseHas('menus', ['name' => 'Menu Statistik']);
     }
 
-    /** @test
-     * Menguji jika parameter menu_type tidak dikirim maka default ke 1
-     */
+    #[Test]
     public function test_store_menu_type_default_ke_satu_jika_null()
     {
         $payload = [
