+++ tests/Feature/OtpControllerTest.php
@@ -44,7 +44,7 @@ class OtpControllerTest extends BaseTestCase
         $this->app->instance(TwoFactorService::class, $this->twoFactorService);
     }
 
-    /** @test */
+    #[Test]
     public function it_shows_otp_settings_page()
     {
         $this->twoFactorService
@@ -65,7 +65,7 @@ class OtpControllerTest extends BaseTestCase
         $response->assertViewHas('twoFactorStatus');
     }
 
-    /** @test */
+    #[Test]
     public function it_shows_otp_activation_page()
     {
         $response = $this->get(route('otp.activate'));
@@ -75,7 +75,7 @@ class OtpControllerTest extends BaseTestCase
         $response->assertViewHas('user', $this->user);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_setup_otp_with_email()
     {
         $this->otpService
@@ -104,7 +104,7 @@ class OtpControllerTest extends BaseTestCase
         $this->assertEquals($this->user->email, session('temp_otp_config.identifier'));
     }
 
-    /** @test */
+    #[Test]
     public function it_can_setup_otp_with_telegram()
     {
         $this->otpService
@@ -133,7 +133,7 @@ class OtpControllerTest extends BaseTestCase
         $this->assertEquals($this->user->telegram_chat_id, session('temp_otp_config.identifier'));
     }
 
-    /** @test */
+    #[Test]
     public function it_handles_otp_setup_failure()
     {
         $this->otpService
@@ -156,7 +156,7 @@ class OtpControllerTest extends BaseTestCase
         ]);
     }
 
-    /** @test */
+    #[Test]
     public function it_enforces_rate_limiting_on_otp_setup()
     {
         // Hit rate limit with new key format (includes IP and User-Agent)
@@ -176,7 +176,7 @@ class OtpControllerTest extends BaseTestCase
         ]);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_verify_otp_activation()
     {
         session(['temp_otp_config' => [
@@ -211,7 +211,7 @@ class OtpControllerTest extends BaseTestCase
         $this->assertArrayNotHasKey('temp_otp_config', session()->all());
     }
 
-    /** @test */
+    #[Test]
     public function it_rejects_otp_verification_without_temp_config()
     {
         $response = $this->postJson(route('otp.verify-activation'), [
@@ -225,7 +225,7 @@ class OtpControllerTest extends BaseTestCase
         ]);
     }
 
-    /** @test */
+    #[Test]
     public function it_handles_otp_verification_failure()
     {
         session(['temp_otp_config' => [
@@ -253,7 +253,7 @@ class OtpControllerTest extends BaseTestCase
         ]);
     }
 
-    /** @test */
+    #[Test]
     public function it_enforces_rate_limiting_on_otp_verification()
     {
         session(['temp_otp_config' => [
@@ -278,7 +278,7 @@ class OtpControllerTest extends BaseTestCase
         ]);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_disable_otp()
     {
         $this->user->update([
@@ -301,7 +301,7 @@ class OtpControllerTest extends BaseTestCase
         $this->assertNull($this->user->otp_identifier);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_resend_otp()
     {
         session(['temp_otp_config' => [
@@ -329,7 +329,7 @@ class OtpControllerTest extends BaseTestCase
         ]);
     }
 
-    /** @test */
+    #[Test]
     public function it_rejects_resend_without_temp_config()
     {
         $response = $this->postJson(route('otp.resend'));
@@ -341,7 +341,7 @@ class OtpControllerTest extends BaseTestCase
         ]);
     }
 
-    /** @test */
+    #[Test]
     public function it_enforces_rate_limiting_on_otp_resend()
     {
         session(['temp_otp_config' => [
@@ -364,7 +364,7 @@ class OtpControllerTest extends BaseTestCase
         ]);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_disable_2fa_from_otp_controller()
     {
         $this->twoFactorService
