+++ tests/Feature/PageControllerWebsiteTest.php
@@ -11,7 +11,7 @@ use Tests\WebsiteTestCase;
 
 class PageControllerWebsiteTest extends WebsiteTestCase
 {
-    /** @test */
+    #[Test]
     public function it_can_access_index_page()
     {
         $response = $this->get(route('web.index'));
@@ -19,7 +19,7 @@ class PageControllerWebsiteTest extends WebsiteTestCase
         $response->assertViewIs('web.index');
     }
 
-    /** @test */
+    #[Test]
     public function it_can_access_category_page()
     {
         $category = Category::factory()->create();
@@ -29,7 +29,7 @@ class PageControllerWebsiteTest extends WebsiteTestCase
         $response->assertViewHas('title', $category->name);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_access_page_detail()
     {
         $page = Page::factory()->create();
@@ -39,7 +39,7 @@ class PageControllerWebsiteTest extends WebsiteTestCase
         $response->assertViewHas('object', $page);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_access_article_detail()
     {
         $article = Article::factory()->create();
@@ -49,7 +49,7 @@ class PageControllerWebsiteTest extends WebsiteTestCase
         $response->assertViewHas('object', $article);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_access_sitemap()
     {
         $mock = Mockery::mock(SitemapService::class);
